refactor: vertically organize BlockFetch
